Move GetCommandsMap where it is used
